Parallel processing (#27)
* Ignoring "Obj" field

"Obj" field refers to the original `FuncDecl` object of method
defination where that function is getting called. This is inside
CallExp. In the event method call occurs before the function declration.
In that case while generating Json `FuncDecl` object gets added inside
`CallExp`. At the place of original `FuncDecl` node, we get the
reference node id.

Hence ignoring this field at the momemnt.

* Concurrency changes

Made changes to run the parser in parallel threads without any errors.
We are limiting the number threads to available number of processors on
the machine.